refactor(dashboard): update main_screen and home_view to use existing files
- Remove new files (main_screen_new.dart, home_view_enhanced.dart, home_view_old.dart)
- Update existing main_screen.dart with TabBar navigation and all components
- Enhance existing home_view.dart with WelcomeCard and QuickActions
- Keep only the original files structure without creating new files
- All dashboard features (Tasks #16-22) now integrated into existing files
