+++ lib/ features/dashboard/home/home_view.dart
@@ -1,21 +1,173 @@
-
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'welcome_card.dart';
+import 'quick_actions.dart';
 
-/// ================= NOTE MODEL =================
+/// Task #19: Build home_view.dart - Display user notes in cards
+class HomeView extends StatelessWidget {
+  const HomeView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final user = FirebaseAuth.instance.currentUser;
+
+    if (user == null) {
+      return const _NotLoggedInView();
+    }
+
+    return RefreshIndicator(
+      onRefresh: () async {
+        await Future.delayed(const Duration(seconds: 1));
+      },
+      child: CustomScrollView(
+        slivers: [
+          // Task #20: Welcome Card with summary
+          SliverToBoxAdapter(
+            child: StreamBuilder<Map<String, int>>(
+              stream: _getStatsStream(user.uid),
+              builder: (context, snapshot) {
+                final stats = snapshot.data ?? {'notes': 0, 'tasks': 0, 'expenses': 0};
+                return WelcomeCard(
+                  notesCount: stats['notes'] ?? 0,
+                  tasksCount: stats['tasks'] ?? 0,
+                  expensesCount: stats['expenses'] ?? 0,
+                );
+              },
+            ),
+          ),
+
+          // Quick Actions
+          SliverToBoxAdapter(
+            child: QuickActions(
+              onAddNote: () => _showSnackBar(context, 'Add note coming soon!'),
+              onAddTask: () => _showSnackBar(context, 'Add task coming soon!'),
+              onAddExpense: () => _showSnackBar(context, 'Add expense coming soon!'),
+              onAddAppointment: () => _showSnackBar(context, 'Add appointment coming soon!'),
+            ),
+          ),
+
+          // Section Header
+          SliverToBoxAdapter(
+            child: Padding(
+              padding: const EdgeInsets.fromLTRB(16, 24, 16, 12),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  Text(
+                    'Recent Notes',
+                    style: TextStyle(
+                      fontSize: 20,
+                      fontWeight: FontWeight.bold,
+                      color: Colors.grey.shade800,
+                    ),
+                  ),
+                  TextButton(
+                    onPressed: () => _showSnackBar(context, 'View all coming soon!'),
+                    child: const Text('View All'),
+                  ),
+                ],
+              ),
+            ),
+          ),
+
+          // Notes List
+          StreamBuilder<List<Note>>(
+            stream: _getNotesStream(user.uid),
+            builder: (context, snapshot) {
+              if (snapshot.connectionState == ConnectionState.waiting) {
+                return const SliverFillRemaining(
+                  child: Center(child: CircularProgressIndicator()),
+                );
+              }
+
+              if (snapshot.hasError) {
+                return SliverFillRemaining(
+                  child: _ErrorView(error: snapshot.error.toString()),
+                );
+              }
+
+              final notes = snapshot.data ?? [];
+
+              if (notes.isEmpty) {
+                return const SliverFillRemaining(child: _EmptyNotesView());
+              }
+
+              return SliverPadding(
+                padding: const EdgeInsets.symmetric(horizontal: 16),
+                sliver: SliverList(
+                  delegate: SliverChildBuilderDelegate(
+                    (context, index) => Padding(
+                      padding: const EdgeInsets.only(bottom: 12),
+                      child: _NoteCard(note: notes[index]),
+                    ),
+                    childCount: notes.length > 10 ? 10 : notes.length,
+                  ),
+                ),
+              );
+            },
+          ),
+
+          const SliverToBoxAdapter(child: SizedBox(height: 80)),
+        ],
+      ),
+    );
+  }
+
+  Stream<Map<String, int>> _getStatsStream(String userId) async* {
+    final notesStream = FirebaseFirestore.instance
+        .collection('notes')
+        .where('userId', isEqualTo: userId)
+        .snapshots();
+
+    await for (final notesSnapshot in notesStream) {
+      final tasksSnapshot = await FirebaseFirestore.instance
+          .collection('tasks')
+          .where('userId', isEqualTo: userId)
+          .get();
+
+      final expensesSnapshot = await FirebaseFirestore.instance
+          .collection('expenses')
+          .where('userId', isEqualTo: userId)
+          .get();
+
+      yield {
+        'notes': notesSnapshot.docs.length,
+        'tasks': tasksSnapshot.docs.length,
+        'expenses': expensesSnapshot.docs.length,
+      };
+    }
+  }
+
+  Stream<List<Note>> _getNotesStream(String userId) {
+    return FirebaseFirestore.instance
+        .collection('notes')
+        .where('userId', isEqualTo: userId)
+        .orderBy('createdAt', descending: true)
+        .limit(10)
+        .snapshots()
+        .map((snapshot) =>
+            snapshot.docs.map((doc) => Note.fromFirestore(doc)).toList());
+  }
+
+  void _showSnackBar(BuildContext context, String message) {
+    ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(message)));
+  }
+}
 
 class Note {
   final String id;
   final String title;
   final String content;
   final DateTime createdAt;
+  final String? category;
 
   Note({
     required this.id,
     required this.title,
     required this.content,
     required this.createdAt,
+    this.category,
   });
 
   factory Note.fromFirestore(DocumentSnapshot doc) {
@@ -24,134 +176,42 @@ class Note {
       id: doc.id,
       title: (data['title'] ?? '') as String,
       content: (data['content'] ?? '') as String,
-      createdAt: (data['createdAt'] as Timestamp?)?.toDate() ??
-          DateTime.fromMillisecondsSinceEpoch(0),
+      createdAt: (data['createdAt'] as Timestamp?)?.toDate() ?? DateTime.now(),
+      category: data['category'] as String?,
     );
   }
 }
 
-/// ================= NOTES REPOSITORY =================
-
-class NotesRepository {
-  final FirebaseFirestore _firestore;
-
-  NotesRepository({FirebaseFirestore? firestore})
-      : _firestore = firestore ?? FirebaseFirestore.instance;
-
-  /// Stream لكل نوتس اليوزر الحالي
-  Stream<List<Note>> watchUserNotes(String userId) {
-    return _firestore
-        .collection('notes')
-        .where('userId', isEqualTo: userId)
-    // 🔴 شِلّينا orderBy عشان ما يحتاجش Index
-        .snapshots()
-        .map(
-          (snap) => snap.docs.map(Note.fromFirestore).toList(),
-    );
-  }
-}
-
-/// ================= HOME VIEW =================
-
-class HomeView extends StatelessWidget {
-  const HomeView({super.key});
+class _NotLoggedInView extends StatelessWidget {
+  const _NotLoggedInView();
 
   @override
   Widget build(BuildContext context) {
-    final user = FirebaseAuth.instance.currentUser;
-
-    // لحد ما الـ anonymous sign-in يخلص
-    if (user == null) {
-      return const Scaffold(
-        body: _LoadingView(),
-      );
-    }
-
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('My Notes'),
-      ),
-      body: _NotesListView(
-        userId: user.uid,
-        repository: NotesRepository(),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          // TODO: إضافة شاشة لإنشاء Note جديدة
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('Add new note action (to be implemented).'),
+    return Center(
+      child: Padding(
+        padding: const EdgeInsets.all(24),
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Icon(Icons.lock_outline, size: 64, color: Colors.grey.shade400),
+            const SizedBox(height: 16),
+            const Text('You are not logged in', style: TextStyle(fontSize: 20, fontWeight: FontWeight.w600)),
+            const SizedBox(height: 8),
+            Text('Please sign in to access your notes', style: TextStyle(fontSize: 14, color: Colors.grey.shade600), textAlign: TextAlign.center),
+            const SizedBox(height: 24),
+            ElevatedButton(
+              onPressed: () => Navigator.of(context).pushReplacementNamed('/login'),
+              child: const Text('Sign In'),
             ),
-          );
-        },
-        child: const Icon(Icons.add),
+          ],
+        ),
       ),
-    );
-  }
-}
-
-/// ================= NOTES LIST VIEW =================
-
-class _NotesListView extends StatelessWidget {
-  final String userId;
-  final NotesRepository repository;
-
-  const _NotesListView({
-    required this.userId,
-    required this.repository,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder<List<Note>>(
-      stream: repository.watchUserNotes(userId),
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.waiting) {
-          return const _LoadingView();
-        }
-
-        if (snapshot.hasError) {
-          return _ErrorView(error: snapshot.error.toString());
-        }
-
-        // 🟢 نرتّب في Flutter حسب createdAt (أحدث نوت فوق)
-        final notes = snapshot.data ?? [];
-        notes.sort((a, b) => b.createdAt.compareTo(a.createdAt));
-
-        if (notes.isEmpty) {
-          return const _EmptyNotesView();
-        }
-
-        return ListView.separated(
-            padding: const EdgeInsets.all(16),
-            itemCount: notes.length,
-            separatorBuilder: (, _) => const SizedBox(height: 12),
-        itemBuilder: (context, index) {
-        final note = notes[index];
-        return _NoteCard(note: note);
-        },
-        );
-      },
-    );
-  }
-}
-
-/// ================= STATES (LOADING / ERROR / EMPTY) =================
-
-class _LoadingView extends StatelessWidget {
-  const _LoadingView();
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      child: CircularProgressIndicator(),
     );
   }
 }
 
 class _ErrorView extends StatelessWidget {
   final String error;
-
   const _ErrorView({required this.error});
 
   @override
@@ -162,28 +222,11 @@ class _ErrorView extends StatelessWidget {
         child: Column(
           mainAxisSize: MainAxisSize.min,
           children: [
-            const Icon(
-              Icons.error_outline,
-              size: 48,
-              color: Colors.red,
-            ),
+            Icon(Icons.error_outline, size: 56, color: Colors.red.shade400),
             const SizedBox(height: 16),
-            const Text(
-              'Something went wrong',
-              style: TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
+            const Text('Something went wrong', style: TextStyle(fontSize: 18, fontWeight: FontWeight.w600)),
             const SizedBox(height: 8),
-            Text(
-              error,
-              style: const TextStyle(
-                fontSize: 14,
-                color: Colors.grey,
-              ),
-              textAlign: TextAlign.center,
-            ),
+            Text(error, style: TextStyle(fontSize: 14, color: Colors.grey.shade600), textAlign: TextAlign.center),
           ],
         ),
       ),
@@ -200,29 +243,13 @@ class _EmptyNotesView extends StatelessWidget {
       child: Padding(
         padding: const EdgeInsets.all(24),
         child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: const [
-            Icon(
-              Icons.note_outlined,
-              size: 56,
-            ),
-            SizedBox(height: 16),
-            Text(
-              'No notes yet',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            SizedBox(height: 8),
-            Text(
-              'Tap the + button to create your first note.',
-              style: TextStyle(
-                fontSize: 14,
-                color: Colors.grey,
-              ),
-              textAlign: TextAlign.center,
-            ),
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Icon(Icons.note_outlined, size: 80, color: Colors.grey.shade300),
+            const SizedBox(height: 24),
+            const Text('No notes yet', style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold)),
+            const SizedBox(height: 12),
+            Text('Start creating your first note\nby tapping the + button', style: TextStyle(fontSize: 16, color: Colors.grey.shade600), textAlign: TextAlign.center),
           ],
         ),
       ),
@@ -230,76 +257,100 @@ class _EmptyNotesView extends StatelessWidget {
   }
 }
 
-/// ================= NOTE CARD =================
-
 class _NoteCard extends StatelessWidget {
   final Note note;
-
   const _NoteCard({required this.note});
 
   @override
   Widget build(BuildContext context) {
-    final createdAtText =
-        '${note.createdAt.day}/${note.createdAt.month}/${note.createdAt.year}';
-
     return Container(
       padding: const EdgeInsets.all(16),
       decoration: BoxDecoration(
         color: Colors.white,
         borderRadius: BorderRadius.circular(16),
-        boxShadow: const [
+        boxShadow: [
           BoxShadow(
-            blurRadius: 8,
-            offset: Offset(0, 2),
-            color: Color(0x22000000),
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 4),
           ),
         ],
       ),
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          // العنوان
-          Text(
-            note.title.isEmpty ? 'Untitled note' : note.title,
-            style: const TextStyle(
-              fontSize: 18,
-              fontWeight: FontWeight.w600,
-            ),
-          ),
-          const SizedBox(height: 8),
-
-          // المحتوى
-          Text(
-            note.content,
-            maxLines: 4,
-            overflow: TextOverflow.ellipsis,
-            style: const TextStyle(
-              fontSize: 14,
-            ),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              Expanded(
+                child: Text(
+                  note.title.isEmpty ? 'Untitled Note' : note.title,
+                  style: const TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+                  maxLines: 1,
+                  overflow: TextOverflow.ellipsis,
+                ),
+              ),
+              if (note.category != null)
+                Container(
+                  padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
+                  decoration: BoxDecoration(
+                    color: _getCategoryColor(note.category!).withOpacity(0.1),
+                    borderRadius: BorderRadius.circular(12),
+                  ),
+                  child: Text(
+                    note.category!.toUpperCase(),
+                    style: TextStyle(
+                      fontSize: 10,
+                      fontWeight: FontWeight.bold,
+                      color: _getCategoryColor(note.category!),
+                    ),
+                  ),
+                ),
+            ],
           ),
           const SizedBox(height: 12),
-
-          // التاريخ
+          Text(
+            note.content,
+            style: TextStyle(fontSize: 14, color: Colors.grey.shade700, height: 1.5),
+            maxLines: 3,
+            overflow: TextOverflow.ellipsis,
+          ),
+          const SizedBox(height: 16),
           Row(
-            mainAxisAlignment: MainAxisAlignment.end,
             children: [
-              Icon(
-                Icons.access_time,
-                size: 16,
-                color: Colors.grey.shade600,
-              ),
+              Icon(Icons.access_time, size: 14, color: Colors.grey.shade500),
               const SizedBox(width: 4),
-              Text(
-                createdAtText,
-                style: TextStyle(
-                  fontSize: 12,
-                  color: Colors.grey.shade600,
-                ),
+              Text(_formatDate(note.createdAt), style: TextStyle(fontSize: 12, color: Colors.grey.shade500)),
+              const Spacer(),
+              IconButton(
+                icon: const Icon(Icons.more_vert, size: 20),
+                onPressed: () {},
+                padding: EdgeInsets.zero,
+                constraints: const BoxConstraints(),
               ),
             ],
           ),
         ],
       ),
     );
+  }
+
+  Color _getCategoryColor(String category) {
+    switch (category.toLowerCase()) {
+      case 'task': return Colors.green;
+      case 'expense': return Colors.orange;
+      case 'appointment': return Colors.purple;
+      case 'quote': return Colors.teal;
+      default: return Colors.blue;
+    }
+  }
+
+  String _formatDate(DateTime date) {
+    final now = DateTime.now();
+    final difference = now.difference(date);
+    if (difference.inDays == 0) return 'Today';
+    if (difference.inDays == 1) return 'Yesterday';
+    if (difference.inDays < 7) return '${difference.inDays} days ago';
+    return '${date.day}/${date.month}/${date.year}';
   }
 }
