--- lib/ features/dashboard/home/home_view_enhanced.dart
@@ -1,494 +1,0 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
-import 'welcome_card.dart';
-import 'stats_grid.dart';
-import 'quick_actions.dart';
-
-/// Home View - Enhanced with Welcome Card, Stats, and Quick Actions
-/// Displays user's recent notes with summary
-class HomeView extends StatelessWidget {
-  const HomeView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final user = FirebaseAuth.instance.currentUser;
-
-    if (user == null) {
-      return const _NotLoggedInView();
-    }
-
-    return RefreshIndicator(
-      onRefresh: () async {
-        // Refresh data
-        await Future.delayed(const Duration(seconds: 1));
-      },
-      child: CustomScrollView(
-        slivers: [
-          // Welcome Card
-          SliverToBoxAdapter(
-            child: StreamBuilder<Map<String, int>>(
-              stream: _getStatsStream(user.uid),
-              builder: (context, snapshot) {
-                final stats = snapshot.data ??
-                    {'notes': 0, 'tasks': 0, 'expenses': 0};
-                return WelcomeCard(
-                  notesCount: stats['notes'] ?? 0,
-                  tasksCount: stats['tasks'] ?? 0,
-                  expensesCount: stats['expenses'] ?? 0,
-                );
-              },
-            ),
-          ),
-
-          // Quick Actions
-          SliverToBoxAdapter(
-            child: QuickActions(
-              onAddNote: () {
-                ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(content: Text('Add note coming soon!')),
-                );
-              },
-              onAddTask: () {
-                ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(content: Text('Add task coming soon!')),
-                );
-              },
-              onAddExpense: () {
-                ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(content: Text('Add expense coming soon!')),
-                );
-              },
-              onAddAppointment: () {
-                ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(content: Text('Add appointment coming soon!')),
-                );
-              },
-            ),
-          ),
-
-          // Section Header
-          SliverToBoxAdapter(
-            child: Padding(
-              padding: const EdgeInsets.fromLTRB(16, 24, 16, 12),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Text(
-                    'Recent Notes',
-                    style: TextStyle(
-                      fontSize: 20,
-                      fontWeight: FontWeight.bold,
-                      color: Colors.grey.shade800,
-                    ),
-                  ),
-                  TextButton(
-                    onPressed: () {
-                      ScaffoldMessenger.of(context).showSnackBar(
-                        const SnackBar(content: Text('View all coming soon!')),
-                      );
-                    },
-                    child: const Text('View All'),
-                  ),
-                ],
-              ),
-            ),
-          ),
-
-          // Notes List
-          StreamBuilder<List<Note>>(
-            stream: _getNotesStream(user.uid),
-            builder: (context, snapshot) {
-              if (snapshot.connectionState == ConnectionState.waiting) {
-                return const SliverFillRemaining(
-                  child: Center(child: CircularProgressIndicator()),
-                );
-              }
-
-              if (snapshot.hasError) {
-                return SliverFillRemaining(
-                  child: _ErrorView(error: snapshot.error.toString()),
-                );
-              }
-
-              final notes = snapshot.data ?? [];
-
-              if (notes.isEmpty) {
-                return const SliverFillRemaining(
-                  child: _EmptyNotesView(),
-                );
-              }
-
-              return SliverPadding(
-                padding: const EdgeInsets.symmetric(horizontal: 16),
-                sliver: SliverList(
-                  delegate: SliverChildBuilderDelegate(
-                    (context, index) {
-                      return Padding(
-                        padding: const EdgeInsets.only(bottom: 12),
-                        child: _NoteCard(note: notes[index]),
-                      );
-                    },
-                    childCount: notes.length > 10 ? 10 : notes.length,
-                  ),
-                ),
-              );
-            },
-          ),
-
-          // Bottom Padding
-          const SliverToBoxAdapter(
-            child: SizedBox(height: 80),
-          ),
-        ],
-      ),
-    );
-  }
-
-  /// Get stats stream for all categories
-  Stream<Map<String, int>> _getStatsStream(String userId) async* {
-    final notesStream = FirebaseFirestore.instance
-        .collection('notes')
-        .where('userId', isEqualTo: userId)
-        .snapshots();
-
-    final tasksStream = FirebaseFirestore.instance
-        .collection('tasks')
-        .where('userId', isEqualTo: userId)
-        .snapshots();
-
-    final expensesStream = FirebaseFirestore.instance
-        .collection('expenses')
-        .where('userId', isEqualTo: userId)
-        .snapshots();
-
-    await for (final notesSnapshot in notesStream) {
-      final tasksSnapshot = await tasksStream.first;
-      final expensesSnapshot = await expensesStream.first;
-
-      yield {
-        'notes': notesSnapshot.docs.length,
-        'tasks': tasksSnapshot.docs.length,
-        'expenses': expensesSnapshot.docs.length,
-      };
-    }
-  }
-
-  /// Get notes stream
-  Stream<List<Note>> _getNotesStream(String userId) {
-    return FirebaseFirestore.instance
-        .collection('notes')
-        .where('userId', isEqualTo: userId)
-        .orderBy('createdAt', descending: true)
-        .limit(10)
-        .snapshots()
-        .map((snapshot) =>
-            snapshot.docs.map((doc) => Note.fromFirestore(doc)).toList());
-  }
-}
-
-/// Note Model
-class Note {
-  final String id;
-  final String title;
-  final String content;
-  final DateTime createdAt;
-  final String? category;
-
-  Note({
-    required this.id,
-    required this.title,
-    required this.content,
-    required this.createdAt,
-    this.category,
-  });
-
-  factory Note.fromFirestore(DocumentSnapshot doc) {
-    final data = doc.data() as Map<String, dynamic>? ?? {};
-    return Note(
-      id: doc.id,
-      title: (data['title'] ?? '') as String,
-      content: (data['content'] ?? '') as String,
-      createdAt: (data['createdAt'] as Timestamp?)?.toDate() ??
-          DateTime.now(),
-      category: data['category'] as String?,
-    );
-  }
-}
-
-/// Not Logged In View
-class _NotLoggedInView extends StatelessWidget {
-  const _NotLoggedInView();
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Icon(
-              Icons.lock_outline,
-              size: 64,
-              color: Colors.grey.shade400,
-            ),
-            const SizedBox(height: 16),
-            const Text(
-              'You are not logged in',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            const SizedBox(height: 8),
-            Text(
-              'Please sign in to access your notes',
-              style: TextStyle(
-                fontSize: 14,
-                color: Colors.grey.shade600,
-              ),
-              textAlign: TextAlign.center,
-            ),
-            const SizedBox(height: 24),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.of(context).pushReplacementNamed('/login');
-              },
-              child: const Text('Sign In'),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-/// Error View
-class _ErrorView extends StatelessWidget {
-  final String error;
-
-  const _ErrorView({required this.error});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Icon(
-              Icons.error_outline,
-              size: 56,
-              color: Colors.red.shade400,
-            ),
-            const SizedBox(height: 16),
-            const Text(
-              'Something went wrong',
-              style: TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            const SizedBox(height: 8),
-            Text(
-              error,
-              style: TextStyle(
-                fontSize: 14,
-                color: Colors.grey.shade600,
-              ),
-              textAlign: TextAlign.center,
-            ),
-            const SizedBox(height: 24),
-            ElevatedButton(
-              onPressed: () {
-                // Reload
-              },
-              child: const Text('Retry'),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-/// Empty Notes View
-class _EmptyNotesView extends StatelessWidget {
-  const _EmptyNotesView();
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              Icons.note_outlined,
-              size: 80,
-              color: Colors.grey.shade300,
-            ),
-            const SizedBox(height: 24),
-            const Text(
-              'No notes yet',
-              style: TextStyle(
-                fontSize: 24,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            const SizedBox(height: 12),
-            Text(
-              'Start creating your first note\nby tapping the + button',
-              style: TextStyle(
-                fontSize: 16,
-                color: Colors.grey.shade600,
-              ),
-              textAlign: TextAlign.center,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-/// Note Card
-class _NoteCard extends StatelessWidget {
-  final Note note;
-
-  const _NoteCard({required this.note});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(16),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.05),
-            blurRadius: 10,
-            offset: const Offset(0, 4),
-          ),
-        ],
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          // Header
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Expanded(
-                child: Text(
-                  note.title.isEmpty ? 'Untitled Note' : note.title,
-                  style: const TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.bold,
-                  ),
-                  maxLines: 1,
-                  overflow: TextOverflow.ellipsis,
-                ),
-              ),
-              if (note.category != null)
-                Container(
-                  padding:
-                      const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
-                  decoration: BoxDecoration(
-                    color: _getCategoryColor(note.category!).withOpacity(0.1),
-                    borderRadius: BorderRadius.circular(12),
-                  ),
-                  child: Text(
-                    note.category!.toUpperCase(),
-                    style: TextStyle(
-                      fontSize: 10,
-                      fontWeight: FontWeight.bold,
-                      color: _getCategoryColor(note.category!),
-                    ),
-                  ),
-                ),
-            ],
-          ),
-
-          const SizedBox(height: 12),
-
-          // Content
-          Text(
-            note.content,
-            style: TextStyle(
-              fontSize: 14,
-              color: Colors.grey.shade700,
-              height: 1.5,
-            ),
-            maxLines: 3,
-            overflow: TextOverflow.ellipsis,
-          ),
-
-          const SizedBox(height: 16),
-
-          // Footer
-          Row(
-            children: [
-              Icon(
-                Icons.access_time,
-                size: 14,
-                color: Colors.grey.shade500,
-              ),
-              const SizedBox(width: 4),
-              Text(
-                _formatDate(note.createdAt),
-                style: TextStyle(
-                  fontSize: 12,
-                  color: Colors.grey.shade500,
-                ),
-              ),
-              const Spacer(),
-              IconButton(
-                icon: const Icon(Icons.more_vert, size: 20),
-                onPressed: () {
-                  // Show options menu
-                },
-                padding: EdgeInsets.zero,
-                constraints: const BoxConstraints(),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
-  Color _getCategoryColor(String category) {
-    switch (category.toLowerCase()) {
-      case 'task':
-        return Colors.green;
-      case 'expense':
-        return Colors.orange;
-      case 'appointment':
-        return Colors.purple;
-      case 'quote':
-        return Colors.teal;
-      default:
-        return Colors.blue;
-    }
-  }
-
-  String _formatDate(DateTime date) {
-    final now = DateTime.now();
-    final difference = now.difference(date);
-
-    if (difference.inDays == 0) {
-      return 'Today';
-    } else if (difference.inDays == 1) {
-      return 'Yesterday';
-    } else if (difference.inDays < 7) {
-      return '${difference.inDays} days ago';
-    } else {
-      return '${date.day}/${date.month}/${date.year}';
-    }
-  }
-}
