--- lib/ features/dashboard/home/home_view_old.dart
@@ -1,305 +1,0 @@
-
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
-
-/// ================= NOTE MODEL =================
-
-class Note {
-  final String id;
-  final String title;
-  final String content;
-  final DateTime createdAt;
-
-  Note({
-    required this.id,
-    required this.title,
-    required this.content,
-    required this.createdAt,
-  });
-
-  factory Note.fromFirestore(DocumentSnapshot doc) {
-    final data = doc.data() as Map<String, dynamic>? ?? {};
-    return Note(
-      id: doc.id,
-      title: (data['title'] ?? '') as String,
-      content: (data['content'] ?? '') as String,
-      createdAt: (data['createdAt'] as Timestamp?)?.toDate() ??
-          DateTime.fromMillisecondsSinceEpoch(0),
-    );
-  }
-}
-
-/// ================= NOTES REPOSITORY =================
-
-class NotesRepository {
-  final FirebaseFirestore _firestore;
-
-  NotesRepository({FirebaseFirestore? firestore})
-      : _firestore = firestore ?? FirebaseFirestore.instance;
-
-  /// Stream لكل نوتس اليوزر الحالي
-  Stream<List<Note>> watchUserNotes(String userId) {
-    return _firestore
-        .collection('notes')
-        .where('userId', isEqualTo: userId)
-    // 🔴 شِلّينا orderBy عشان ما يحتاجش Index
-        .snapshots()
-        .map(
-          (snap) => snap.docs.map(Note.fromFirestore).toList(),
-    );
-  }
-}
-
-/// ================= HOME VIEW =================
-
-class HomeView extends StatelessWidget {
-  const HomeView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final user = FirebaseAuth.instance.currentUser;
-
-    // لحد ما الـ anonymous sign-in يخلص
-    if (user == null) {
-      return const Scaffold(
-        body: _LoadingView(),
-      );
-    }
-
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('My Notes'),
-      ),
-      body: _NotesListView(
-        userId: user.uid,
-        repository: NotesRepository(),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          // TODO: إضافة شاشة لإنشاء Note جديدة
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('Add new note action (to be implemented).'),
-            ),
-          );
-        },
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
-}
-
-/// ================= NOTES LIST VIEW =================
-
-class _NotesListView extends StatelessWidget {
-  final String userId;
-  final NotesRepository repository;
-
-  const _NotesListView({
-    required this.userId,
-    required this.repository,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder<List<Note>>(
-      stream: repository.watchUserNotes(userId),
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.waiting) {
-          return const _LoadingView();
-        }
-
-        if (snapshot.hasError) {
-          return _ErrorView(error: snapshot.error.toString());
-        }
-
-        // 🟢 نرتّب في Flutter حسب createdAt (أحدث نوت فوق)
-        final notes = snapshot.data ?? [];
-        notes.sort((a, b) => b.createdAt.compareTo(a.createdAt));
-
-        if (notes.isEmpty) {
-          return const _EmptyNotesView();
-        }
-
-        return ListView.separated(
-            padding: const EdgeInsets.all(16),
-            itemCount: notes.length,
-            separatorBuilder: (, _) => const SizedBox(height: 12),
-        itemBuilder: (context, index) {
-        final note = notes[index];
-        return _NoteCard(note: note);
-        },
-        );
-      },
-    );
-  }
-}
-
-/// ================= STATES (LOADING / ERROR / EMPTY) =================
-
-class _LoadingView extends StatelessWidget {
-  const _LoadingView();
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      child: CircularProgressIndicator(),
-    );
-  }
-}
-
-class _ErrorView extends StatelessWidget {
-  final String error;
-
-  const _ErrorView({required this.error});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const Icon(
-              Icons.error_outline,
-              size: 48,
-              color: Colors.red,
-            ),
-            const SizedBox(height: 16),
-            const Text(
-              'Something went wrong',
-              style: TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            const SizedBox(height: 8),
-            Text(
-              error,
-              style: const TextStyle(
-                fontSize: 14,
-                color: Colors.grey,
-              ),
-              textAlign: TextAlign.center,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _EmptyNotesView extends StatelessWidget {
-  const _EmptyNotesView();
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: const [
-            Icon(
-              Icons.note_outlined,
-              size: 56,
-            ),
-            SizedBox(height: 16),
-            Text(
-              'No notes yet',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            SizedBox(height: 8),
-            Text(
-              'Tap the + button to create your first note.',
-              style: TextStyle(
-                fontSize: 14,
-                color: Colors.grey,
-              ),
-              textAlign: TextAlign.center,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-/// ================= NOTE CARD =================
-
-class _NoteCard extends StatelessWidget {
-  final Note note;
-
-  const _NoteCard({required this.note});
-
-  @override
-  Widget build(BuildContext context) {
-    final createdAtText =
-        '${note.createdAt.day}/${note.createdAt.month}/${note.createdAt.year}';
-
-    return Container(
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(16),
-        boxShadow: const [
-          BoxShadow(
-            blurRadius: 8,
-            offset: Offset(0, 2),
-            color: Color(0x22000000),
-          ),
-        ],
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          // العنوان
-          Text(
-            note.title.isEmpty ? 'Untitled note' : note.title,
-            style: const TextStyle(
-              fontSize: 18,
-              fontWeight: FontWeight.w600,
-            ),
-          ),
-          const SizedBox(height: 8),
-
-          // المحتوى
-          Text(
-            note.content,
-            maxLines: 4,
-            overflow: TextOverflow.ellipsis,
-            style: const TextStyle(
-              fontSize: 14,
-            ),
-          ),
-          const SizedBox(height: 12),
-
-          // التاريخ
-          Row(
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: [
-              Icon(
-                Icons.access_time,
-                size: 16,
-                color: Colors.grey.shade600,
-              ),
-              const SizedBox(width: 4),
-              Text(
-                createdAtText,
-                style: TextStyle(
-                  fontSize: 12,
-                  color: Colors.grey.shade600,
-                ),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
