+++ lib/ features/dashboard/main_screen.dart
@@ -1,339 +1,115 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 
-/// ================= NOTE MODEL =================
+import 'navigation/app_drawer.dart';
+import 'navigation/tab_bar.dart';
+import 'home/home_view.dart';
+import 'categories/tasks_view.dart';
+import 'categories/expenses_view.dart';
+import 'categories/appointments_view.dart';
+import 'categories/quotes_view.dart';
+import 'search/search_functionality.dart';
 
-class Note {
-  final String id;
-  final String title;
-  final String content;
-  final DateTime createdAt;
+class MainScreen extends StatefulWidget {
+  const MainScreen({super.key});
 
-  Note({
-    required this.id,
-    required this.title,
-    required this.content,
-    required this.createdAt,
-  });
-
-  factory Note.fromFirestore(DocumentSnapshot doc) {
-    final data = doc.data() as Map<String, dynamic>? ?? {};
-    return Note(
-      id: doc.id,
-      title: (data['title'] ?? '') as String,
-      content: (data['content'] ?? '') as String,
-      createdAt: (data['createdAt'] as Timestamp?)?.toDate() ??
-          DateTime.fromMillisecondsSinceEpoch(0),
-    );
-  }
+  @override
+  State<MainScreen> createState() => _MainScreenState();
 }
 
-/// ================= NOTES REPOSITORY =================
+class _MainScreenState extends State<MainScreen> with SingleTickerProviderStateMixin {
+  late TabController _tabController;
+  int _currentIndex = 0;
 
-class NotesRepository {
-  final FirebaseFirestore _firestore;
-
-  NotesRepository({FirebaseFirestore? firestore})
-      : _firestore = firestore ?? FirebaseFirestore.instance;
-
-  /// Stream لكل نوتس اليوزر الحالي
-  Stream<List<Note>> watchUserNotes(String userId) {
-    return _firestore
-        .collection('notes')
-        .where('userId', isEqualTo: userId)
-        .orderBy('createdAt', descending: true)
-        .snapshots()
-        .map(
-          (snap) => snap.docs.map(Note.fromFirestore).toList(),
-    );
+  @override
+  void initState() {
+    super.initState();
+    _tabController = TabController(length: 5, vsync: this);
+    _tabController.addListener(() {
+      setState(() => _currentIndex = _tabController.index);
+    });
   }
-}
 
-/// ================= HOME VIEW =================
-
-class HomeView extends StatelessWidget {
-  const HomeView({super.key});
+  @override
+  void dispose() {
+    _tabController.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
     final user = FirebaseAuth.instance.currentUser;
+    
+    if (user == null) {
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        Navigator.of(context).pushReplacementNamed('/login');
+      });
+      return const Scaffold(body: Center(child: CircularProgressIndicator()));
+    }
 
     return Scaffold(
       appBar: AppBar(
-        title: const Text('My Notes'),
-      ),
-      body: user == null
-          ? const _NotLoggedInView()
-          : _NotesListView(
-        userId: user.uid,
-        repository: NotesRepository(),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          // TODO: إضافة شاشة لإنشاء Note جديدة
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('Add new note action (to be implemented).'),
+        title: Text(_getAppBarTitle()),
+        actions: [
+          IconButton(
+            icon: const Icon(Icons.search),
+            onPressed: () => Navigator.push(
+              context,
+              MaterialPageRoute(builder: (context) => const SearchFunctionality()),
             ),
-          );
-        },
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
-}
-
-/// ================= NOT LOGGED IN VIEW =================
-
-class _NotLoggedInView extends StatelessWidget {
-  const _NotLoggedInView();
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const Icon(
-              Icons.lock_outline,
-              size: 56,
+          ),
+          IconButton(
+            icon: const Icon(Icons.notifications_outlined),
+            onPressed: () => ScaffoldMessenger.of(context).showSnackBar(
+              const SnackBar(content: Text('Notifications coming soon!')),
             ),
-            const SizedBox(height: 16),
-            const Text(
-              'You are not logged in',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            const SizedBox(height: 8),
-            Text(
-              'Sign in to see your notes from Firebase.',
-              style: TextStyle(
-                fontSize: 14,
-                color: Colors.grey.shade700,
-              ),
-              textAlign: TextAlign.center,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-/// ================= NOTES LIST VIEW =================
-
-class _NotesListView extends StatelessWidget {
-  final String userId;
-  final NotesRepository repository;
-
-  const _NotesListView({
-    required this.userId,
-    required this.repository,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder<List<Note>>(
-      stream: repository.watchUserNotes(userId),
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.waiting) {
-          return const _LoadingView();
-        }
-
-        if (snapshot.hasError) {
-          return _ErrorView(error: snapshot.error.toString());
-        }
-
-        final notes = snapshot.data ?? [];
-
-        if (notes.isEmpty) {
-          return const _EmptyNotesView();
-        }
-
-        return ListView.separated(
-            padding: const EdgeInsets.all(16),
-            itemCount: notes.length,
-            separatorBuilder: (, _) => const SizedBox(height: 12),
-        itemBuilder: (context, index) {
-        final note = notes[index];
-        return _NoteCard(note: note);
-        },
-        );
-      },
-    );
-  }
-}
-
-/// ================= STATES (LOADING / ERROR / EMPTY) =================
-
-class _LoadingView extends StatelessWidget {
-  const _LoadingView();
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      child: CircularProgressIndicator(),
-    );
-  }
-}
-
-class _ErrorView extends StatelessWidget {
-  final String error;
-
-  const _ErrorView({required this.error});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const Icon(
-              Icons.error_outline,
-              size: 48,
-              color: Colors.red,
-            ),
-            const SizedBox(height: 16),
-            const Text(
-              'Something went wrong',
-              style: TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            const SizedBox(height: 8),
-            Text(
-              error,
-              style: TextStyle(
-                fontSize: 14,
-                color: Colors.grey.shade700,
-              ),
-              textAlign: TextAlign.center,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _EmptyNotesView extends StatelessWidget {
-  const _EmptyNotesView();
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const Icon(
-              Icons.note_outlined,
-              size: 56,
-            ),
-            const SizedBox(height: 16),
-            const Text(
-              'No notes yet',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            const SizedBox(height: 8),
-            Text(
-              'Tap the + button to create your first note.',
-              style: TextStyle(
-                fontSize: 14,
-                color: Colors.grey.shade700,
-              ),
-              textAlign: TextAlign.center,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-/// ================= NOTE CARD =================
-
-class _NoteCard extends StatelessWidget {
-  final Note note;
-
-  const _NoteCard({required this.note});
-
-  @override
-  Widget build(BuildContext context) {
-    final createdAtText =
-        '${note.createdAt.day}/${note.createdAt.month}/${note.createdAt.year}';
-
-    return Container(
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(16),
-        boxShadow: const [
-          BoxShadow(
-            blurRadius: 8,
-            offset: Offset(0, 2),
-            color: Color(0x22000000),
           ),
         ],
       ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
+      drawer: const AppDrawer(),
+      body: Column(
         children: [
-          // العنوان
-          Text(
-            note.title.isEmpty ? 'Untitled note' : note.title,
-            style: const TextStyle(
-              fontSize: 18,
-              fontWeight: FontWeight.w600,
+          DashboardTabBar(tabController: _tabController),
+          Expanded(
+            child: TabBarView(
+              controller: _tabController,
+              children: const [
+                HomeView(),
+                TasksView(),
+                ExpensesView(),
+                AppointmentsView(),
+                QuotesView(),
+              ],
             ),
-          ),
-          const SizedBox(height: 8),
-
-          // المحتوى
-          Text(
-            note.content,
-            maxLines: 4,
-            overflow: TextOverflow.ellipsis,
-            style: TextStyle(
-              fontSize: 14,
-              color: Colors.grey.shade800,
-            ),
-          ),
-          const SizedBox(height: 12),
-
-          // التاريخ
-          Row(
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: [
-              Icon(
-                Icons.access_time,
-                size: 16,
-                color: Colors.grey.shade600,
-              ),
-              const SizedBox(width: 4),
-              Text(
-                createdAtText,
-                style: TextStyle(
-                  fontSize: 12,
-                  color: Colors.grey.shade600,
-                ),
-              ),
-            ],
           ),
         ],
       ),
+      floatingActionButton: _buildFAB(),
+    );
+  }
+
+  String _getAppBarTitle() {
+    const titles = ['Nota - Home', 'Tasks', 'Expenses', 'Appointments', 'Quotes'];
+    return titles[_currentIndex];
+  }
+
+  Widget _buildFAB() {
+    final actions = [
+      (Icons.add, 'Add Note'),
+      (Icons.add_task, 'Add Task'),
+      (Icons.add, 'Add Expense'),
+      (Icons.event, 'Add Appointment'),
+      (Icons.format_quote, 'Add Quote'),
+    ];
+    
+    final (icon, tooltip) = actions[_currentIndex];
+    
+    return FloatingActionButton(
+      onPressed: () => ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(content: Text(' coming soon!')),
+      ),
+      tooltip: tooltip,
+      child: Icon(icon),
     );
   }
 }
